chore: Collapse runner packages
