+++ internal/runner/builder.go
@@ -1,29 +1,28 @@
-package runnerpool
+package runner
 
 import (
 	"context"
 
-	"github.com/gruntwork-io/terragrunt/internal/runner/common"
 	"github.com/gruntwork-io/terragrunt/internal/venv"
 	"github.com/gruntwork-io/terragrunt/pkg/log"
 	"github.com/gruntwork-io/terragrunt/pkg/options"
 )
 
-// Build stack runner using discovery and queueing mechanisms. v is the
-// virtualized environment used by the per-unit version-constraint probe.
-func Build(
+// New discovers all Terragrunt units under the working directory and assembles them
+// into a [Runner] that can apply or destroy them.
+func New(
 	ctx context.Context,
 	l log.Logger,
 	v *venv.Venv,
 	opts *options.TerragruntOptions,
-	runnerOpts ...common.Option,
-) (common.StackRunner, error) {
+	runnerOpts ...Option,
+) (*Runner, error) {
 	discovered, err := discoverWithRetry(ctx, l, v, opts, runnerOpts...)
 	if err != nil {
 		return nil, err
 	}
 
-	rnr, err := createRunner(ctx, l, opts, discovered, runnerOpts...)
+	rnr, err := createRunner(ctx, l, opts, discovered)
 	if err != nil {
 		return nil, err
 	}
