--- internal/runner/runnerpool/export_test.go
@@ -1,4 +1,0 @@
-package runnerpool
-
-// MaxDependencyTraversalDepth exposes the dependency traversal bound to tests.
-const MaxDependencyTraversalDepth = maxDependencyTraversalDepth
